add caching to the Tool base class.
Expensive initilizations can contribute to slowdowns in large toolboxes as each tool subclass has its initilizer called on toolbox load and toolbox refresh.

Each tool will also re-initilize when opened, and during update operations to parameters. So storing the result of an init into a static cache will actually bring the behavior of the tool more in line with a standard Python object